+++ src/domains/admin/admin-dashboard/admin-dashboard-match-stats/api/react-query-api/use-get-match-stats.tsx
@@ -6,10 +6,15 @@ import { getMatchStatsData } from "../admin-dashboard-match-stats-api";
 import { adminDashboardMatchStatsKeys } from "./admin-dashboard-match-stats-query-keys";
 import { useSuspenseQuery } from "@tanstack/react-query";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export const useGetMatchStats = () => {
   const { data } = useSuspenseQuery({
     queryKey: adminDashboardMatchStatsKeys.stats(),
-    queryFn: getMatchStatsData,
+    queryFn: async () => {
+      const response = await getMatchStatsData();
+      return handleSupabaseApiResponse(response);
+    },
     staleTime: 1000 * 60 * 5, // 5분
   });