+++ src/domains/admin/admin-match/admin-match-lineup/api/react-query-api/use-get-match-lineups-suspense.tsx
@@ -3,15 +3,15 @@ import { useSuspenseQuery } from "@tanstack/react-query";
 
 import { getMatchLineups } from "@admin/admin-match/admin-match-lineup/api/admin-match-lineup-api";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export const useGetMatchLineupsSuspense = (matchId: string) => {
   return useSuspenseQuery({
     queryKey: ADMIN_MATCH_LINEUP_QUERY_KEYS.list(matchId),
     queryFn: async () => {
       const response = await getMatchLineups(matchId);
-      if (response.error) {
-        throw new Error(response.error.message);
-      }
-      return response.data || [];
+
+      return handleSupabaseApiResponse(response);
     },
   });
 };