+++ src/domains/admin/admin-match/api/react-query-api/use-get-all-matches-suspense.tsx
@@ -3,15 +3,14 @@ import { useSuspenseQuery } from "@tanstack/react-query";
 
 import { getAllMatches } from "@admin/admin-match/api/admin-match-api";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export const useGetAllMatchesSuspense = () => {
   return useSuspenseQuery({
     queryKey: ADMIN_MATCH_QUERY_KEYS.lists(),
     queryFn: async () => {
       const response = await getAllMatches();
-      if (response.error) {
-        throw new Error(response.error.message);
-      }
-      return response.data || [];
+      return handleSupabaseApiResponse(response);
     },
   });
 };