+++ src/domains/admin/admin-team/api/react-query-api/use-create-team.tsx
@@ -3,16 +3,15 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { type ICreateTeamRequest, createTeam } from "@admin/admin-team/api/admin-team-api";
 import { ADMIN_TEAM_QUERY_KEYS } from "@admin/admin-team/api/react-query-api/admin-team-query-keys";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export function useCreateTeam() {
   const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn: async (team: ICreateTeamRequest) => {
       const response = await createTeam(team);
-      if (response.error) {
-        throw new Error(response.error.message);
-      }
-      return response.data;
+      return handleSupabaseApiResponse(response, team);
     },
     onSuccess: () => {
       queryClient.invalidateQueries({