+++ src/domains/admin/admin-user/api/react-query-api/use-delete-user.tsx
@@ -2,11 +2,16 @@ import { deleteUserByAdmin } from "../admin-user-api";
 import { adminUserQueryKeys } from "./admin-user-query-keys";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export const useDeleteUser = () => {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: deleteUserByAdmin,
+    mutationFn: async (userId: string) => {
+      const response = await deleteUserByAdmin(userId);
+      return handleSupabaseApiResponse(response, userId);
+    },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: adminUserQueryKeys.users() });
     },