+++ src/domains/admin/admin-user/api/react-query-api/use-get-all-users-suspense.tsx
@@ -2,15 +2,14 @@ import { getAllUsers } from "../admin-user-api";
 import { adminUserQueryKeys } from "./admin-user-query-keys";
 import { useSuspenseQuery } from "@tanstack/react-query";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export const useGetAllUsersSuspense = () => {
   return useSuspenseQuery({
     queryKey: adminUserQueryKeys.users(),
     queryFn: async () => {
       const response = await getAllUsers();
-      if (response.error) {
-        throw new Error(response.error.message);
-      }
-      return response.data;
+      return handleSupabaseApiResponse(response);
     },
   });
 };