+++ src/domains/animals/api/react-query-api/animalQueryKeys.ts
@@ -1,3 +1,0 @@
-export const animalQueryKeys = {
-  animals: "animals",
-};