+++ src/domains/animals/mocks/animal-handlers.ts
@@ -1,8 +1,0 @@
-import { HttpResponse, http } from "msw";
-import AnimalDummy from "./animal-dummy.json";
-
-export const AnimalHandlers = [
-  http.get("*/animal", () => {
-    return HttpResponse.json(AnimalDummy);
-  }),
-];