+++ src/domains/matches/matches-history/matches-history-players-rating/api/react-query-api/use-get-matches-history-players-rating-suspense.tsx
@@ -6,6 +6,8 @@ import {
 } from "@matches/matches-history/matches-history-players-rating/api/matches-history-players-rating-api";
 import { MATCHES_HISTORY_PLAYERS_RATING_QUERY_KEYS } from "@matches/matches-history/matches-history-players-rating/api/react-query-api/matches-history-players-rating-query-keys";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export function useGetMatchesHistoryPlayersRatingSuspense(matchId: string) {
   const results = useSuspenseQueries({
     queries: [
@@ -13,16 +15,14 @@ export function useGetMatchesHistoryPlayersRatingSuspense(matchId: string) {
         queryKey: [MATCHES_HISTORY_PLAYERS_RATING_QUERY_KEYS.MATCHES_HISTORY_PLAYERS_RATING, matchId],
         queryFn: async () => {
           const response = await getMatchesHistoryPlayersRating(matchId);
-          if (response.error) throw new Error(response.error.message);
-          return response.data || [];
+          return handleSupabaseApiResponse(response);
         },
       },
       {
         queryKey: [MATCHES_HISTORY_PLAYERS_RATING_QUERY_KEYS.MATCHES_HISTORY_PLAYERS_RATING, matchId, "info"],
         queryFn: async () => {
           const response = await getMatchInfo(matchId);
-          if (response.error) throw new Error(response.error.message);
-          return response.data;
+          return handleSupabaseApiResponse(response);
         },
       },
     ],