+++ src/domains/players/players-rating-rotator/api/react-query-api/use-get-players-rating-rotator-acc.tsx
@@ -6,10 +6,15 @@ import {
 } from "@players/players-rating-rotator/api/players-rating-rotator-api";
 import { PLAYER_RATING_ROTATOR_QUERY_KEYS } from "@players/players-rating-rotator/api/react-query-api/players-rating-rotator-query-keys";
 
+import { handleSupabaseApiResponse } from "@shared/utils/sentry-utils";
+
 export function useGetPlayersRatingRotatorAcc(): IRotatePlayerStatAccumulated[] {
   const { data } = useSuspenseQuery<IRotatePlayerStatAccumulated[]>({
     queryKey: [PLAYER_RATING_ROTATOR_QUERY_KEYS.PLAYER_RATING_ROTATOR_ACC],
-    queryFn: getPlayersRatingRotatorAcc,
+    queryFn: async () => {
+      const response = await getPlayersRatingRotatorAcc();
+      return handleSupabaseApiResponse(response);
+    },
   });
 
   return data;